pyproject.toml
[project]
name = "poc-voice-agent"
version = "0.1.0"
description = "PoC for VoiceAgent"
readme = "README.md"
requires-python = ">=3.13.5"
dependencies = [
    "faster-whisper>=1.2.0",
    "librosa>=0.11.0",
    "llama-cpp-python>=0.3.16",
    "loguru>=0.7.3",
    "numpy>=1.26.0",
    "openwakeword>=0.4.0",
    "pandas>=2.0.0",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "rich>=14.1.0",
    "scipy>=1.11.0",
    "sentencepiece>=0.2.1",
    "sounddevice>=0.5.2",
    "soundfile>=0.13.1",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "transformers>=4.56.0",
    "webrtcvad>=2.0.10",
]
